requirement.txt
python-dotenv
os
json
requests
torch
fastapi
transformers
pydantic
streamlit